introduce irc shard: * move main part of responsibilities to IrcShard * add but NOT use MessageBuckets
